app/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
av==15.0.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.0
faster-whisper==1.1.1
ffmpeg==1.4
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
google-auth==2.40.3
google-genai==1.24.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
jira==3.8.0
jiter==0.10.0
mpmath==1.3.0
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.93.1
packaging==25.0
pillow==11.3.0
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyreadline3==3.5.4
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tenacity==8.5.0
tokenizers==0.21.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1